Fixed algorithm errors and changed selection/mutation methods
Fixed documentation in functions.

Reordered functions to make more sense of the code.

Changed selection method to Roulette Selection.

Changed algorithm to use the global best individual in newer generations.

Removed all the parents references in each individual as it leads to heap out of memory.

Fixed error where the algorithm was removing cities from the route. This happened because the first best individual was empty
at the start of the execution.

Changed mutation function as the current method mutates the population more aggressively and the fitness is more instable.
The new method mutates the population more passively and the fitness is more stable. It also runs faster as it doesn't loop over
all the cities of each individual.

Updated ReadMe.