Fixed mutation and selection functions
Selection now uses elitism and random roulette selection instead of just roulette selection.
After mutating, the algorithm was not updating the fitness and total distance values of the
individuals. This resulted in incorrect best individuals and results.
Mutation now has a random probability of swapping all the cities of an individual instead
of just swapping one city.
The algorithm improved from finding the best result in 10000 generations to just 200 generations.